funcion incremento y decremento implementada
