+++ src/components/ItemCount/ItemCount.js
@@ -1,39 +1,18 @@
-import React, { useState } from 'react'
+const ItemCount = ({sumarClick, restarClick, contador}) => {
 
-const ItemCount = ({title}) => {
-
-    const [click, setClick] = useState(0);
-
-    //Funcion incremento de clicks
-    const addClick = () =>{
-        if(click > 9){
-            alert(`cantidad Limitada`);
-        }else{
-            setClick(click + 1);
-        };
-    };
-
-    //Funcion decremento de clicks
-    const removeClick = () => {
-        if(click > 0){
-            setClick (click - 1);
-        }else{
-            alert(`Seleccione un producto`)
-        }
-    }
 
     return (
 
         <>
             <div>
-                <h1>{title}</h1>
+                <h1>{contador}</h1>
             </div>
             <div>
                 <p></p>
             </div>
             <div>
-                <button className="count_button">+</button>
-                <button className="count_button">-</button>
+                <button className="count_button" onClick={sumarClick}>+</button>
+                <button className="count_button" onClick={restarClick}>-</button>
             </div>
